+++ src/utilities/testing/mocking/MockedProvider.tsx
@@ -15,7 +15,7 @@ export interface MockedProviderProps<TSerializedCache = {}> {
   cache?: ApolloCache<TSerializedCache>;
   resolvers?: Resolvers;
   childProps?: object;
-  children?: React.ReactElement;
+  children?: any;
   link?: ApolloLink;
 }
 
@@ -57,7 +57,7 @@ export class MockedProvider extends React.Component<
 
   public render() {
     const { children, childProps } = this.props;
-    return children ? (
+    return React.isValidElement(children) ? (
       <ApolloProvider client={this.state.client}>
         {React.cloneElement(React.Children.only(children), { ...childProps })}
       </ApolloProvider>