+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## Apollo Client 3.3.12 (not yet released)
+## Apollo Client 3.3.12
 
 ### Bug fixes